+++ language/german.php
@@ -450,6 +450,8 @@ $LANG_SHOP = array (
 'back_to_reports' => 'Back to report selection',
 'config_report' => 'Re-configure this report',
 'no_exp' => 'No Expiration',
+'price_incl_disc' => 'Price reflects any applicable discounts.',
+'reflects_disc' => 'Price reflects a %s%% discount.',
 );
 if (isset($_SHOP_CONF['ena_ratings']) && $_SHOP_CONF['ena_ratings']) {
     $LANG_SHOP['list_sort_options']['top_rated'] = 'Top Rated';