+++ CHANGELOG.md
@@ -19,3 +19,4 @@ Release TBD
   * Add service function for plugins to send gift cards.
   * Return no search results if shop is not publicly enabled.
   * Fix quantity discount price if items are added to the cart to meet the qty.
+  * Indicate that unit prices in the cart include discounts, if applicable.